Moved helseatlas code from pages/ to src/
Avoid repeating code between english and norwegian pages
